+++ src/mddj/_types.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+import datetime
+import typing as t
+
+TomlValue: t.TypeAlias = (
+    "dict[str, TomlValue] | list[TomlValue] | str | int | datetime.datetime"
+)